Turn password generator into a python module
