+++ .gitignore
@@ -1,0 +1,4 @@
+dist/
+*.egg-info
+*.pyc
+__pycache__/
