+++ setup.cfg
@@ -1,0 +1,19 @@
+[metadata]
+name = password_generator
+version = 1.0.0
+description = Interface graphique qui génère un mot de passe
+url = https://...
+author = Arthur Lasnier
+author_email = arthur.lasnier77700@gmail.com
+maintainer = Arthur Lasnier
+maintainer_email = arthur.lasnier77700@gmail.com
+
+
+[options]
+include_package_data = True
+packages = find:
+package_dir =
+    =src
+
+[options.packages.find]
+where = src
