+++ src/password_generator/__main__.py
@@ -1,0 +1,3 @@
+from password_generator.generator_window import *
+
+app()
