+++ src/password_generator/logo_A.ico
